+++ src/form_helper.js
@@ -3,7 +3,6 @@ import "core-js/fn/array/filter"
 import "core-js/fn/array/find"
 import "core-js/fn/array/from"
 import "core-js/fn/object/values"
-
 export default class FormHelper {
   constructor(widgetConfig, formHelperConfig){
     this.widgetConfig = widgetConfig
